move errorMessages from exception to constants
