+++ textdb/textdb-dataflow/src/main/java/edu/uci/ics/textdb/dataflow/common/AbstractSingleInputOperator.java
@@ -1,8 +1,8 @@
 package edu.uci.ics.textdb.dataflow.common;
 
+import edu.uci.ics.textdb.api.constants.ErrorMessages;
 import edu.uci.ics.textdb.api.dataflow.IOperator;
 import edu.uci.ics.textdb.api.exception.DataFlowException;
-import edu.uci.ics.textdb.api.exception.ErrorMessages;
 import edu.uci.ics.textdb.api.exception.TextDBException;
 import edu.uci.ics.textdb.api.schema.Schema;
 import edu.uci.ics.textdb.api.tuple.Tuple;
