+++ textdb/textdb-storage/src/main/java/edu/uci/ics/textdb/storage/DataWriter.java
@@ -15,8 +15,8 @@ import org.apache.lucene.search.Query;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.FSDirectory;
 
+import edu.uci.ics.textdb.api.constants.ErrorMessages;
 import edu.uci.ics.textdb.api.constants.SchemaConstants;
-import edu.uci.ics.textdb.api.exception.ErrorMessages;
 import edu.uci.ics.textdb.api.exception.StorageException;
 import edu.uci.ics.textdb.api.field.IDField;
 import edu.uci.ics.textdb.api.field.IField;
